# Use orjson instead of json, when available