@JukkaL 2024-10-15
Sounds very promising! I can also perform some measurements. 

> should we make it a required dependency?

I wonder how well maintained orjson is, and does it ship binary wheels for all the platforms we care about? We might be adding ARM Linux wheels at some point, and it would be nice if all our dependencies would ship with binary wheels (though it's perhaps not essential for Linux, as long as there are x86-64 wheels).

> I don't know if the sqlite cache path is used at all (what's the status?),

Sqlite caching is very much used, and I'm thinking of enabling it by default in the future. In certain use cases it's significantly faster than a file-per-module cache, and we use it at work.