@hauntsaninja 2024-10-15
I think orjson does ship wheels for all platforms we care about. It would be nice if Python packaging had a concept of a "default" extra