@JukkaL 2024-10-15
I'm seeing a 10-15% improvement to the performance of `time mypy -c 'import torch'` on Linux. This probably also helps incremental mypy runs in general.