@JukkaL 2024-10-16
> Is just using files2 in sqlite a sufficient migration

This seems fine. It's an internal implementation detail, and caches aren't compatible between mypy versions.

Can you check if `misc/convert-cache.py` still works?