@hauntsaninja 2024-10-16
Thanks, there was a missing spot where I'd forgotten to change to files2.

Note that convert-cache.py seems to only half fill the cache, even on master.
```
time python -m mypy --config-file mypy_self_check.ini -p mypy
python misc/convert-cache.py --to-sqlite .mypy_cache --output_dir .mypy_cache/3.8
time python -m mypy --config-file mypy_self_check.ini -p mypy --sqlite-cache  # slow, table definitely found and has entries
time python -m mypy --config-file mypy_self_check.ini -p mypy --sqlite-cache  # fast, table has double the entries
```
I'll look into that separately, since it's an issue on master