@hauntsaninja 2024-10-16
Okay, fixed the cache convert problem on master in https://github.com/python/mypy/pull/17974